Move style handling to style.go
